Created an interface for WorkerQueue and renamed the implementation file
